# src/public/swagger/Auth.yaml
openapi: 3.0.0
info:
  title: School Management API
  version: 1.0.0
  description: API documentation for the School Management System, including authentication and core entities like Users, Kelas, Siswa, and Guru.
...
paths:
...
  /api/v1/users:
    get:
      summary: Get all users
      tags: [Users]
      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new user (for admin use, different from register)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/kelas:
    get:
      summary: Get all kelas
      tags: [Kelas]
      responses:
        200:
          description: A list of kelas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListKelasResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new kelas
      tags: [Kelas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Kelas'
      responses:
        201:
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KelasResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/jurusan:
    get:
      summary: Get all jurusan
      tags: [Jurusan]
      responses:
        200:
          description: A list of jurusan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJurusanResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new jurusan
      tags: [Jurusan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jurusan'
      responses:
        201:
          description: Major created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurusanResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/mata-pelajaran:
    get:
      summary: Get all mata-pelajaran
      tags: [Mata Pelajaran]
      responses:
        200:
          description: A list of mata-pelajaran
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMataPelajaranResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new mata-pelajaran
      tags: [Mata Pelajaran]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MataPelajaran'
      responses:
        201:
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MataPelajaranResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/guru:
    get:
      summary: Get all guru
      tags: [Guru]
      responses:
        200:
          description: A list of guru
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListGuruResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new guru
      tags: [Guru]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Guru'
      responses:
        201:
          description: Teacher created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuruResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/jadwal:
    get:
      summary: Get all jadwal
      tags: [Jadwal]
      responses:
        200:
          description: A list of jadwal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJadwalResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new jadwal
      tags: [Jadwal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jadwal'
      responses:
        201:
          description: Schedule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JadwalResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          description: Indicates whether the request was successful
          example: false
        message:
          type: string
          description: A descriptive error message
          example: An error occurred
      required:
        - status
        - message

    # --- Core Entity Schemas (based on Prisma) ---
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: Unique identifier for the user.
          example: 1
        username:
          type: string
          description: Unique username for the user.
          example: "siswa001"
        name:
          type: string
          description: Full name of the user.
          example: "Budi Santoso"
        role:
          type: string
          enum: [admin, guru, siswa]
          description: Role of the user.
          example: siswa
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Date and time when the user was created.
          example: "2024-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Date and time when the user was last updated.
          example: "2024-01-01T10:00:00Z"
    CreateUser:
      type: object
      required:
        - username
        - name
        - role
        - password
      properties:
        username:
          type: string
          description: Unique username for the user.
          example: "admin01"
        name:
          type: string
          description: Full name of the user.
          example: "Admin System"
        role:
          type: string
          enum: [admin, guru, siswa]
          description: Role of the user.
          example: admin
        password:
          type: string
          format: password
          minLength: 8
          description: Password for the user account.
          example: "strongpassword"
...
    UserResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: User fetched successfully
        data:
          $ref: '#/components/schemas/User'
    ListUsersResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Users fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    Kelas:
      type: object
      required:
        - nama_kelas
        - kelas_romawi
      properties:
        id_kelas:
          type: integer
          readOnly: true
          description: Unique identifier for the kelas.
          example: 1
        nama_kelas:
          type: string
          description: Name of the kelas (e.g., "10 IPA 1").
          example: "10 IPA 1"
        kelas_romawi:
          type: string
          description: Roman numeral representation of the kelas level (e.g., "X", "XI", "XII").
          example: "X"
      example:
        nama_kelas: "10 IPA 1"
        kelas_romawi: "X"
    KelasResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Class fetched successfully
        data:
          $ref: '#/components/schemas/Kelas'
    ListKelasResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Kelas fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Kelas'
    Jurusan:
      type: object
      required:
        - nama_jurusan
        - deskripsi
      properties:
        id_jurusan:
          type: integer
          readOnly: true
          description: Unique identifier for the jurusan.
          example: 1
        nama_jurusan:
          type: string
          description: Name of the jurusan (e.g., "IPA", "IPS", "Bahasa").
          example: "IPA"
        deskripsi:
          type: string
          description: Description of the jurusan.
          example: "Ilmu Pengetahuan Alam"
      example:
        nama_jurusan: "IPS"
        deskripsi: "Ilmu Pengetahuan Sosial"
    JurusanResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Major fetched successfully
        data:
          $ref: '#/components/schemas/Jurusan'
    ListJurusanResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Jurusan fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Jurusan'
    MataPelajaran:
      type: object
      required:
        - nama_mapel
      properties:
        id_mapel:
          type: integer
          readOnly: true
          description: Unique identifier for the mata-pelajaran.
          example: 1
        nama_mapel:
          type: string
          description: Name of the mata-pelajaran (e.g., "Matematika", "Fisika", "Bahasa Indonesia").
          example: "Matematika"
      example:
        nama_mapel: "Fisika"
    MataPelajaranResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Subject fetched successfully
        data:
          $ref: '#/components/schemas/MataPelajaran'
    ListMataPelajaranResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Mata Pelajaran fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/MataPelajaran'
    Guru:
      type: object
      required:
        - nip
        - nama_guru
        - id_mapel
      properties:
        id_guru:
          type: integer
          readOnly: true
          description: Unique identifier for the guru.
          example: 1
        nip:
          type: string
          description: National Educator Identification Number (Nomor Induk Pendidik) for the guru.
          example: "198001012005011001"
        nama_guru:
          type: string
          description: Full name of the guru.
          example: "Dra. Siti Aminah"
        id_mapel:
          type: integer
          description: ID of the mata-pelajaran this guru teaches.
          example: 101
      example:
        nip: "198001012005011001"
        nama_guru: "Dra. Siti Aminah"
        id_mapel: 1
    GuruResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Teacher fetched successfully
        data:
          $ref: '#/components/schemas/Guru'
    ListGuruResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Guru fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Guru'
...
    Jadwal:
      type: object
      required:
        - id_kelas
        - id_mapel
        - id_guru
        - hari
        - jam_mulai
        - jam_selesai
      properties:
        id_jadwal:
          type: integer
          readOnly: true
          description: Unique identifier for the jadwal entry.
          example: 1
        id_kelas:
          type: integer
          description: ID of the kelas for this jadwal.
          example: 1
        id_mapel:
          type: integer
          description: ID of the mata-pelajaran for this jadwal.
          example: 101
        id_guru:
          type: integer
          description: ID of the guru for this jadwal.
          example: 201
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat]
          description: Day of the week for the jadwal.
          example: "Senin"
        jam_mulai:
          type: string
          format: time
          description: Start time of the kelas (HH:MM:SS format).
          example: "07:30:00"
        jam_selesai:
          type: string
          format: time
          description: End time of the kelas (HH:MM:SS format).
          example: "09:00:00"
      example:
        id_kelas: 1
        id_mapel: 1
        id_guru: 1
        hari: "Selasa"
        jam_mulai: "09:15:00"
        jam_selesai: "10:45:00"
    JadwalResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Schedule fetched successfully
        data:
          $ref: '#/components/schemas/Jadwal'
    ListJadwalResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Jadwal fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Jadwal'
...